+++ work_track/src/pages/BusinessView.js
@@ -5,8 +5,8 @@ import CreatePosition from '../components/CreatePosition'
 import EmployeeInfo from '../components/EmployeeInfo'
 import Entries from './Entries'
 import CreateEntry from '../components/CreateEntry'
-import UpdateEmployee from '../components/UpdateDeleteEmployee'
-import UpdatePosition from '../components/UpdateDeletePosition'
+import UpdateEmployee from '../components/UpdateEmployee'
+import UpdatePosition from '../components/UpdatePosition'
 import axios from 'axios'
 
 const BusinessView = ({ props, user, authenticated }) => {
@@ -62,8 +62,8 @@ const BusinessView = ({ props, user, authenticated }) => {
         <CreateEmployee ownerId={ownerId} businessId={businessId} />
         <CreatePosition ownerId={ownerId} businessId={businessId} />
         <CreateEntry ownerId={ownerId} businessId={businessId} logId={logId} />
-        <UpdateEmployee employeeId={employeeId} businessId={businessId} />
-        <UpdatePosition jobId={jobId} businessId={businessId} />
+        {/* <UpdateEmployee employeeId={employeeId} businessId={businessId} /> */}
+        {/* <UpdatePosition jobId={jobId} businessId={businessId} />  */}
       </div>
     </div>
   ) : (
